+++ src/LuaFunction.hpp
@@ -28,6 +28,8 @@ using namespace godot;
 
 namespace luagdextension {
 
+class VariantArguments;
+
 class LuaFunction : public LuaObjectSubclass<sol::protected_function> {
 	GDCLASS(LuaFunction, LuaObject);
 
@@ -39,9 +41,7 @@ public:
 	Variant invokev(const Array& args);
 	Variant invoke(const Variant **args, GDExtensionInt arg_count, GDExtensionCallError &error);
 
-	static Variant invoke_lua(const sol::protected_function& f, const Variant **args, GDExtensionInt arg_count, bool return_lua_error);
-	static Variant invoke_method_lua(const sol::protected_function& f, const Variant& self, const Variant **args, GDExtensionInt arg_count, bool return_lua_error);
-	static Variant invokev_lua(const sol::protected_function& f, const Array& args, bool return_lua_error);
+	static Variant invoke_lua(const sol::protected_function& f, const VariantArguments& args, bool return_lua_error);
 
 	Callable to_callable() const;